chore: rename did integration test file
